requirements.txt
aioboto3==14.1.0
aiobotocore==2.21.1
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.17
aioitertools==0.12.0
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
awsiotsdk
boto3==1.37.1
botocore==1.37.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
fastapi==0.115.12
frozenlist==1.6.0
greenlet==3.2.0
h11==0.14.0
idna==3.10
jmespath==1.0.1
multidict==6.4.3
numpy==2.2.5
opencv-python==4.11.0.86
propcache==0.3.1
pydantic==2.11.3
pydantic_core==2.33.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
requests==2.32.3
s3transfer==0.11.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==1.26.15
uvicorn==0.34.2
wrapt==1.17.2
yarl==1.20.0
